+++ main.py
@@ -1,5 +1,5 @@
 import numpy as np
-
+from custom_plots import GeneratePlots
 # Setup time interval
 fs = 100 # Hz
 time_interval = 10 # sec
@@ -17,15 +17,8 @@ noise = np.random.normal(0, 1, N)
 # Combine signals
 combined_signal = signal_1 + signal_2 + signal_3 + noise
 
-def fft(signal, fs):
-    amp = np.abs(np.fft.fft(signal))
-    freq = np.fft.fftfreq(len(signal), 1/fs)
-    l = int(len(freq) / 2)
-    amp, freq = amp[:l], freq[:l]
-    plt.plot(freq, 20*np.log10(amp))
-    plt.xlabel('Frequency (Hz)'), plt.ylabel('Amplitude (dB)')
-    plt.title('FFT'), plt.grid(), plt.show()
+# Plot
+plotting_obj = GeneratePlots(t, combined_signal)
+plotting_obj.time_plot(), plotting_obj.freq_plot(freq_range=[0, 10])
 
-import matplotlib.pyplot as plt
-plt.plot(t, combined_signal), plt.show()
-fft(combined_signal, fs)
+# source .venv/Scripts/activate
